LayoutTests/http/tests/security/xssAuditor/script-tag-with-trailing-comment2-expected.txt

CONSOLE MESSAGE: line 5: Refused to execute a JavaScript script. Source code of script found within request.
